Modified the navbar
I removed the old "static" navbar and made the sticky navbar permanent
so that it's just always there. To make up for this permanence, I added
some empty space at the top of the webpage for the navbar to occupy in
it's initial state to make sure that no elements are hidden BEHIND the
navbar when it's at the top of the page. I then added some javascript
functionality that makes it so that the current page is automatically
highlited in the navbar, so no more hard coding required, YAY!